Hide showcase when alpha value is 0.00
To make the showcase background color affect the whole page, padding is
used. However, since the padding spans the entire height of the page,
the introduction component can't be selected or highlighted. When the
alpha value is "0.00", padding is swapped out for margin to ensure that
elements below the showcase component can be selected while it's hidden.